splitting Utility
